Annual Income Data Type Change To Decimal and few other Fluent Validation Code Changes

+++ VVPSMS.Api.Domain/Models/FamilyOrGuardianInfoDetail.cs
@@ -21,7 +21,7 @@ public partial class FamilyOrGuardianInfoDetail
 
     public string? DesignationNameofcompany { get; set; }
 
-    public int AnnualIncome { get; set; }
+    public decimal AnnualIncome { get; set; }
 
     public string? OfficeAddress { get; set; }
